requirements.txt
streamlit>=1.30.0
tensorflow>=2.16.0
opencv-python-headless>=4.9.0
numpy
pillow
